reinstalled reactstrap, replaced old navbar with newer one
