+++ app/frontend/robin/package.json
@@ -18,6 +18,7 @@
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.0.4",
     "react-star-rating-component": "^1.4.1",
+    "reactstrap": "^6.5.0",
     "redux": "^4.0.0",
     "uuid": "^3.3.2"
   },
