+++ app/frontend/robin/src/pages/eventpage/components/navbar/index.css
@@ -1,9 +1,10 @@
 .nav {
-  background-color: LightGrey   ;
+  background-color: #f7f7f7;
   -moz-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
   -webkit-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
   box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
+  margin-bottom: 10px;
 }
-.nav-brand{
-  color:white;
+.options-container{
+  right: 0px;
 }
