+++ app/frontend/robin/src/pages/homepage/components/navbar/index.css
@@ -1,30 +1,10 @@
 .nav {
-  background-color: LightGrey   ;
+  background-color: #f7f7f7;
   -moz-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
   -webkit-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
   box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
   margin-bottom: 10px;
 }
-.nav-brand{
-  color:white;
-}
-.stylish-input-group .input-group-addon{
-  background: white !important; 
-}
-.stylish-input-group .form-control{
-  border-right:0; 
-  box-shadow:0 0 0; 
-  border-color:#ccc;
-}
-.stylish-input-group button{
-  border:0;
-  background:transparent;
-}
-.search-container{
-  width: 45%;
-  max-width: 45%;
-  margin-left: 25%;
-}
-.search-button{
-  max-height: 30px;
+.options-container{
+  right: 0px;
 }
