--- app/frontend/robin/src/pages/homepage/components/navbar2/index.css
@@ -1,10 +1,0 @@
-.nav {
-  background-color: #f7f7f7;
-  -moz-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-  -webkit-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-  box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-  margin-bottom: 10px;
-}
-.options-container{
-  right: 0px;
-}
