+++ app/frontend/robin/src/pages/homepage/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link, Redirect } from "react-router-dom";
 
-import NavBar from "./components/navbar2/index"
+import NavBar from "./components/navbar/index"
 import EventComp from "./components/event/index"
 import borisPhoto from "./boris.png"
 import skyGif from "./T9L0.gif"
